Create a checkout of the dotfiles in $HOME
Create a .git file in the $HOME directory, and checkout and reset to set the
state of the git repository to a sane state.

We clone a bare repository, since we don't want to have an actual checkout of
the files in ~/code/my-repos/dotfiles/. But, core.bare flag needs to be set to
false, with $HOME/.git preset to identify it as a checkout of this repo.